requirements.txt
aiohttp
requests
pytest
pytest-asyncio
tenacity
matplotlib
azure-identity